+++ kernel/cpu/idt.h
@@ -18,7 +18,7 @@ typedef struct idt_ptr {
 
 extern void load_idt(idt_ptr *ptr);
 
-void idt_set_gate(uint8_t num, uint32_t base, uint16_t sel, uint8_t flags);
+void idt_set_gate(int num, uint32_t base, uint16_t sel, uint8_t flags);
 
 void idt_init(void);
